pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "oil-gas-well-segmentation"
version = "1.0.0"
description = "Multispectal satellite image segmentation for oil/gas well detection using ResUNet + PyTorch DDP"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Lucas Iturriaga", email = "" }]
license = { text = "MIT" }

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "hydra-core>=1.3.2",
    "pydantic>=2.5.0",
    "webdataset>=0.2.86",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "albumentations>=1.4.0",
    "Pillow>=10.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
